+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -21,6 +21,7 @@
 ## Upcoming Release
 * Introduced NetworkAclBypass and NetworkAclBypassResourceIds for Database Account cmdlets.
 * Introduced ServerVersion option to Update-AzCosmosDBAccount.
+* Introduced BackupInterval and BackupRetention for Database Account cmdlets
 
 ## Version 1.0.0
 * General availability of 'Az.CosmosDB' module